Can now generate video tags without generating frame input

+++ video.py
@@ -37,7 +37,7 @@ from PIL import Image
 parser = argparse.ArgumentParser(description='Process input data for training a Sequence Processor.')
 parser.add_argument('builddir', help='build (timestamp only) that is to be tested')
 parser.add_argument('-v', '--video', default='v', help='video that is to be processed')
-parser.add_argument('-s', '--skipvids', default=False, action='store_true', help='do not extract the frames from the video again')
+parser.add_argument('-s', '--skipvids', default=False, action='store_true', help='do not extract the frames from the video')
 parser.add_argument('-i', '--iter', type=int, default=0, help='use a specific model iteration')
 parser.add_argument('-f', '--framerate', type=float, default=1.0, help='how many frames to store and process per second')
 parser.add_argument('-c', '--confidence-threshold', type=float, default=0.20, help='keep detections with confidence above threshold')
@@ -49,6 +49,12 @@ folder_input = os.path.join(os.getcwd(), 'video', 'input')
 folder_output = os.path.join(os.getcwd(), 'video', 'output')
 if not os.path.exists(folder_output):
     os.makedirs(folder_output)
+folder_tags = os.path.join(folder_output,'tags')
+if not os.path.exists(folder_tags):
+    os.makedirs(folder_tags)
+folder_resolution = os.path.join(folder_output,'resolution')
+if not os.path.exists(folder_resolution):
+    os.makedirs(folder_resolution)
 
 
 # gets the most recent iteration for a certain model build
@@ -86,49 +92,62 @@ shutil.copy2(os.path.join('builds', args.builddir, 'snapshots', recentmodel), 'd
 
 
 # setup DeepDetect service if necessary
-dd = DD('localhost')
-dd.set_return_format(dd.RETURN_PYTHON)
-model = {'repository':'/dockershare/ssd/dedemodel'}
-parameters_input = {'connector':'image', 'width':512, 'height':512}
-parameters_mllib = {'nclasses':7}
-parameters_output = {}
-detect = dd.delete_service('ssd')
-detect = dd.put_service('ssd', model, 'single-shot detector', 'caffe', parameters_input, parameters_mllib, parameters_output, 'supervised')
+if args.skipvids == False:
+    dd = DD('localhost')
+    dd.set_return_format(dd.RETURN_PYTHON)
+    model = {'repository':'/dockershare/ssd/dedemodel'}
+    parameters_input = {'connector':'image', 'width':512, 'height':512}
+    parameters_mllib = {'nclasses':7}
+    parameters_output = {}
+    detect = dd.delete_service('ssd')
+    detect = dd.put_service('ssd', model, 'single-shot detector', 'caffe', parameters_input, parameters_mllib, parameters_output, 'supervised')
 
 
 # assert the txt files directly, just to be sure they are readable
-print ('Asserting all txt files in input folder')
+print ('Converting tags txt to tags csv')
 for root, dirs, files in os.walk(folder_input):
     for name in files:
         name, ext = os.path.splitext(name)
-        if (ext.lower().endswith(('.mp4', '.avi', '.mov')) and os.path.exists(os.path.join(root,name+'.txt'))):
+        if ext.lower().endswith('.txt'):
             with open(os.path.join(root,name+'.txt')) as txt:
-                # process per line
-                for line in txt:
-                    line = line.replace(' ', '')
-                    if (len(line.strip()) != 0):
-                        # translate human format to machine format
-                        line = line.replace('nodig', '0')
-                        line = line.replace('drive', '0')
-                        line = line.replace('rotate', '0')
-                        line = line.replace('lowerraise', '0')
-                        line = line.replace('standstill', '0')
-                        line = line.replace('notinscene', '0')
-                        line = line.replace('dig', '1')
-                        line = line.replace('unusable', '2')
-                        line = line.replace(':', '')
-                        # if the assertion fails, the tag file contains an error
-                        assert len(line.strip()) == 5
+                with open(os.path.join(folder_tags,name+'.csv'), 'w+') as csv:
+                    # empty csv file
+                    csv.truncate()
+                    # process per line
+                    for line in txt:
+                        line = line.replace(' ', '')
+                        if (len(line.strip()) != 0):
+                            # translate human format to machine format
+                            line = line.replace('nodig', '0')
+                            line = line.replace('drive', '0')
+                            line = line.replace('rotate', '0')
+                            line = line.replace('lowerraise', '0')
+                            line = line.replace('standstill', '0')
+                            line = line.replace('notinscene', '0')
+                            line = line.replace('dig', '1')
+                            line = line.replace('unusable', '2')
+                            line = line.replace(':', '')
+                            # if the assertion fails, the tag file contains an error
+                            assert len(line.strip()) == 5
+                            # grab from line
+                            minutes = line[0:2]
+                            seconds = line[2:4]
+                            status = line[4:5]
+                            # calculate and push a new csv line
+                            pushlength = (int(seconds) + int(minutes) * 60) + 1
+                            csv.write(str(pushlength)+','+str(status))
+                            csv.write('\n')
 
 
-# recursively process directory
+# recursively process input directory
 for root, dirs, files in os.walk(folder_input):
     for name in files:
         name, ext = os.path.splitext(name)
         if (
             ext.lower().endswith(('.mp4', '.avi', '.mov')) and
             os.path.exists(os.path.join(root,name+'.txt')) and
-            (args.video == 'v' or args.video == name)
+            (args.video == 'v' or args.video == name) and
+            args.skipvids == False
         ):
 
 
@@ -146,26 +165,18 @@ for root, dirs, files in os.walk(folder_input):
             output_json = os.path.join(folder_output,'json',name)
             if not os.path.exists(output_json):
                 os.makedirs(output_json)
-            folder_tags = os.path.join(folder_output,'tags')
-            if not os.path.exists(folder_tags):
-                os.makedirs(folder_tags)
-            folder_resolution = os.path.join(folder_output,'resolution')
-            if not os.path.exists(folder_resolution):
-                os.makedirs(folder_resolution)
-            output_tags = os.path.join(folder_tags,name+'.csv')
             output_resolution = os.path.join(folder_resolution,name+'.csv')
 
 
             # video -> jpg_unannotated
-            if (args.skipvids == False):
-                print ('  Converting video into unannotated frames...')
-                cmd = 'ffprobe -v error -show_entries format=duration -of default=noprint_wrappers=1:nokey=1 "'+root+'/'+name+ext+'"'
-                duration = os.popen(cmd).read()
-                i = 0.5
-                while (i < float(duration)):
-                    cmd = 'ffmpeg -y -nostats -loglevel 0 -accurate_seek -ss '+str(int(i)/3600).zfill(2)+':'+str(int(i)/60).zfill(2)+':'+str(int(i)%60).zfill(2)+'.5 -t 00:00:01 -i "'+root+'/'+name+ext+'" -r 1 -f singlejpeg "'+output_jpg_unannotated+'/'+name+'_'+str(int(i)+1).zfill(4)+'.jpg"'
-                    os.system(cmd)
-                    i = i + 1
+            print ('  Converting video into unannotated frames...')
+            cmd = 'ffprobe -v error -show_entries format=duration -of default=noprint_wrappers=1:nokey=1 "'+root+'/'+name+ext+'"'
+            duration = os.popen(cmd).read()
+            i = 0.5
+            while (i < float(duration)):
+                cmd = 'ffmpeg -y -nostats -loglevel 0 -accurate_seek -ss '+str(int(i)/3600).zfill(2)+':'+str(int(i)/60).zfill(2)+':'+str(int(i)%60).zfill(2)+'.5 -t 00:00:01 -i "'+root+'/'+name+ext+'" -r 1 -f singlejpeg "'+output_jpg_unannotated+'/'+name+'_'+str(int(i)+1).zfill(4)+'.jpg"'
+                os.system(cmd)
+                i = i + 1
 
 
             # jpg_unannotated -> json, jpg_annotated
@@ -202,35 +213,3 @@ for root, dirs, files in os.walk(folder_input):
             image = Image.open(os.path.join(output_jpg_unannotated,name+'_0001.jpg'))
             with open(output_resolution, 'w+') as res:
                 res.write(str(image.size[0])+','+str(image.size[1])+'\n')
-
-
-            # txt -> csv
-            print ('  Converting tagging txt into csv...')
-            with open(os.path.join(root,name+'.txt')) as txt:
-                with open(output_tags, 'w+') as csv:
-                    # empty csv file
-                    csv.truncate()
-                    # process per line
-                    for line in txt:
-                        line = line.replace(' ', '')
-                        if (len(line.strip()) != 0):
-                            # translate human format to machine format
-                            line = line.replace('nodig', '0')
-                            line = line.replace('drive', '0')
-                            line = line.replace('rotate', '0')
-                            line = line.replace('lowerraise', '0')
-                            line = line.replace('standstill', '0')
-                            line = line.replace('notinscene', '0')
-                            line = line.replace('dig', '1')
-                            line = line.replace('unusable', '2')
-                            line = line.replace(':', '')
-                            # if the assertion fails, the tag file contains an error
-                            assert len(line.strip()) == 5
-                            # grab from line
-                            minutes = line[0:2]
-                            seconds = line[2:4]
-                            status = line[4:5]
-                            # calculate and push a new csv line
-                            pushlength = (int(seconds) + int(minutes) * 60) + 1
-                            csv.write(str(pushlength)+','+str(status))
-                            csv.write('\n')
